+++ common/src/main/kotlin/com/cobblemon/mod/common/api/spawning/CobblemonSpawnPools.kt
@@ -11,6 +11,7 @@ package com.cobblemon.mod.common.api.spawning
 import com.cobblemon.mod.common.api.spawning.condition.BucketPrecalculation
 import com.cobblemon.mod.common.api.spawning.condition.ContextPrecalculation
 import com.cobblemon.mod.common.api.spawning.detail.SpawnPool
+import com.cobblemon.mod.common.data.CobblemonDataProvider
 
 /**
  * A collection of all of Cobblemon's general-purpose [SpawnPool]s. These
@@ -25,6 +26,6 @@ object CobblemonSpawnPools {
     lateinit var WORLD_SPAWN_POOL: SpawnPool
 
     fun load() {
-        WORLD_SPAWN_POOL = SpawnLoader.load("world").addPrecalculators(ContextPrecalculation, BucketPrecalculation)
+        WORLD_SPAWN_POOL = CobblemonDataProvider.register(SpawnPool("world").addPrecalculators(ContextPrecalculation, BucketPrecalculation))
     }
 }